+++ ide_image_scripts/src/entrypoint.sh
@@ -8,15 +8,17 @@ set -e
 
 # source any additional scripts with environment variables
 for SCRIPT in /etc/ide.d/variables/* ; do
-	if [ -f $SCRIPT -a -x $SCRIPT ] ; then
+	if [ -f $SCRIPT ] ; then
+		chmod +x $SCRIPT
 		source $SCRIPT
 	fi
 done
 # run any additional scripts to setup custom configuration files or secrets
 # or source any files or wait for linux daemons
 for SCRIPT in /etc/ide.d/scripts/* ; do
-	if [ -f $SCRIPT -a -x $SCRIPT ] ; then
-		$SCRIPT
+	if [ -f $SCRIPT ] ; then
+		chmod +x $SCRIPT
+		source $SCRIPT
 	fi
 done