+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 0.7.0 (7 Mar 2017)
+
+* \#10674 even if any file in /etc/ide.d/ directory was not executable,
+ make it executable in entrypoint.sh. Because user may forget to make a script
+ executable.
+
 ### 0.6.2 (22 Nov 2016)
 
 * \#10078 fix a bug when validating docker-compose v2 yaml file